pyproject.toml
[tool.poetry]
name = "analysis-ml-template"
version = "0.1.0"
description = ""
authors = ["杉本階嗣"]

[tool.poetry.dependencies]
python = "=3.9.2"
pandas = "^1.2.4"
matplotlib = "^3.4.2"
japanize-matplotlib = "^1.1.3"
seaborn = "^0.11.1"
scikit-learn = "^0.24.2"
python-dotenv = "^0.17.1"
tqdm = "^4.61.0"
pandas-gbq = "^0.15.0"
sweetviz = "^2.1.2"
jupyterlab = "^3.0.16"
hydra-core = "^1.1.1"
PyYAML = "^5.4.1"
Jinja2 = "^3.0.1"
numpy = "^1.23.3"
lightgbm = "^3.3.2"
scipy = "^1.9.1"
google-cloud-storage = "^2.5.0"
jpholiday = "^0.1.8"
google-cloud-aiplatform = "^1.17.0"
timeout-decorator = "^0.5.0"
hydra-colorlog = "^1.2.0"
gcsfs = "^2022.8.2"
invoke = "^1.7.1"
papermill = "^2.4.0"
ipywidgets = "^8.0.2"
google-cloud-logging = "^3.2.5"
kfp = "2.0.0b12"


[tool.poetry.group.dev.dependencies]
black = "22.3"
mypy = "^1.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
